SFA-1478: Base_rate vardef fix.
- Fix the vardefs to make base_rate be a decimal field and also make the default length for decimal fields be 26,6

+++ sugarcrm/modules/ForecastManagerWorksheets/vardefs.php
@@ -70,7 +70,7 @@ $dictionary['ForecastManagerWorksheet'] = array(
         array(
             'name' => 'base_rate',
             'vname' => 'LBL_BASE_RATE',
-            'type' => 'double',
+            'type' => 'decimal',
         ),
         'timeperiod_id' =>
         array(
